+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next
 - `DateInput`: Avoid errors when `undefined` is passed as value
+- Add `isDateDisabled` property
 
 ## 2.6.0 - 2023 Jul 11
 - Add 'yy' format (@erezarnon)