+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Next
+- Add `dynamicPositioning` prop to avoid the date popup appearing outside the screen (@stinger567)
+
 ## 2.6.0 - 2023 Jul 11
 - Add 'yy' format (@erezarnon)